+++ api/v1alpha1/webhook_suite_test.go
@@ -139,6 +139,9 @@ var _ = BeforeSuite(func() {
 	err = (&MySQL{}).SetupWebhookWithManager(mgr)
 	Expect(err).NotTo(HaveOccurred())
 
+	err = (&Cassandra{}).SetupWebhookWithManager(mgr)
+	Expect(err).NotTo(HaveOccurred())
+
 	//+kubebuilder:scaffold:webhook
 
 	go func() {