+++ main.go
@@ -20,7 +20,7 @@ import (
 	//+kubebuilder:scaffold:imports
 )
 
-//go:generate go run ./userconfigs_generator/... --services mysql
+//go:generate go run ./userconfigs_generator/... --services mysql,cassandra
 
 var (
 	scheme   = runtime.NewScheme()
@@ -321,6 +321,19 @@ func main() {
 		os.Exit(1)
 	}
 
+	if err = (&controllers.CassandraReconciler{
+		Controller: controllers.Controller{
+			Client:       mgr.GetClient(),
+			Log:          ctrl.Log.WithName("controllers").WithName("Cassandra"),
+			Scheme:       mgr.GetScheme(),
+			Recorder:     mgr.GetEventRecorderFor("cassandra-reconciler"),
+			DefaultToken: defaultToken,
+		},
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Cassandra")
+		os.Exit(1)
+	}
+
 	if os.Getenv("ENABLE_WEBHOOKS") != "false" {
 		if err = (&v1alpha1.Project{}).SetupWebhookWithManager(mgr); err != nil {
 			setupLog.Error(err, "unable to create webhook", "webhook", "Project")
@@ -406,6 +419,10 @@ func main() {
 		}
 		if err = (&v1alpha1.MySQL{}).SetupWebhookWithManager(mgr); err != nil {
 			setupLog.Error(err, "unable to create webhook", "webhook", "MySQL")
+			os.Exit(1)
+		}
+		if err = (&v1alpha1.Cassandra{}).SetupWebhookWithManager(mgr); err != nil {
+			setupLog.Error(err, "unable to create webhook", "webhook", "Cassandra")
 			os.Exit(1)
 		}
 	}