+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 - Add Cassandra Kind
+- Add Grafana Kind
 
 ## v0.6.0 - 2023-01-16