+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 - Add Cassandra Kind
 - Add Grafana Kind
+- Recreate Kafka ACL if modified. 
+  Note: Modification of ACL created prior to v0.5.1 won't delete existing instance at Aiven.
+  It must be deleted manually.
 
 ## v0.6.0 - 2023-01-16